tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/001.html

<!DOCTYPE html>
<html><head>
  <title> scheduler: inline in markup </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>

  <div id="log">FAILED (This TC requires JavaScript enabled)</div>

  <script>log('inline script #1');</script>
  <script>log('inline script #2');</script>

  <script type="text/javascript">

  var t = async_test()


  function test() {
    assert_array_equals(eventOrder, ['inline script #1', 'inline script #2' ]);
    t.done();
  }
  onload = t.step_func(test)
  </script>

</body></html>
